Added "group" parameter to step 13 (pre.cluster)
Carina's experience was that step 13 failed to multi-thread when the group file was not provided - this made the pre.cluster command take much too long. If mothur is restarted at any point between step 6 and 13 then the groups file will not be part of the default settings and will need to be provided here.